Replaced Sharebar plugin integration
Replaced plugin integration with embedded use of the CSS and JS from the
plugin.  Instead of using the plugin screens (that generate lots of WP
errors), we have a new widget area and you can put whatever you want in
there.  Settings for display have been added to the customizer