Making GitHub Updater Required
It has been over 4 months since I released a version to WordPress.org
and UU2014 still isn't visible in the directory.  I will continue to
deal with that process but I am now making GitHub Updater a "required"
plugin for the theme.  If you don't have it, the theme will tell you to
install and activate it.