Release 1.2.0
### v1.2.0
* One of the most significant cosmetic changes since release

- In response to a request I received, I've updated the theme to be a
bit cleaner and less "rectangular".  Essentially, the background of the
primary sidebar now is the same color as the main content dashboard.
They background to the left and right of the header has also been
converted into a semi-transparent gradient and the edge of the main
content now appears to float above the page background.