Page title updates
Added new page template that will made the title of a page not appear
visually.  It is still there for search engines though.  Also added the
same entry-meta section that posts have to all pages but it is not
visible on pages.  This is also for search engines because they now look
for authors and dates in the semantic HTML for "article".